fix quant qwen convert

+++ neural_speed/convert/convert_quantized_qwen.py
@@ -170,7 +170,7 @@ def main(args_in: Optional[List[str]] = None) -> None:
     if hparams['model_type']=='qwen2':
         fout.write(struct.pack("i", hparams["intermediate_size"]))
     else:
-        fout.write(struct.pack("i", hparams["intermediate_size"]/2))
+        fout.write(struct.pack("i", int(hparams["intermediate_size"]/2)))
     f.write(struct.pack("i", 0))
     f.write(struct.pack("i", 0))  # n_experts
     f.write(struct.pack("i", 0))  # n_expert_used
